Detect bad search results and handle them appropriately
It seems that GitHub's code search changed at some point and it's no
longer possible to get exact file name matches. Generally, the results
rank repos that actually contain a datapackage.json first, but there
may be some mixed in with later results, which is why the script is
conservative and only aborts after 5 full pages of useless results.