Add validate script
For use in conjunction with the lists in this repo, or any other list
of URLs.

The utility iterates through URLs, passes them to the validator API and
returns whether they passed or failed. For failures, a link to the
validator is provided by default, to see why the failure occurred.